jpredapir: A package for interfacing with JPred server.

The jpredapir package provides functions to submit, check status, and retrieve results from JPred: A Secondary Structure Prediction Server (JPred).

Details

This package contains a collection of functions to interface with JPred server.

For a complete list, use library(help = "jpredapir").

The jpredapir package functions

jpredapir::submit() Submit job to JPred REST interface.

jpredapir::status() Check job status on JPred REST interface.

jpredapir::get_results() Download results from JPred server.

jpredapir::quota() Check JPred REST interface quota.

jpredapir::check_rest_version() Check version of JPred REST interface.

References

Browse source code at: https://github.com/MoseleyBioinformaticsLab/jpredapir

Report a bugs at: https://github.com/MoseleyBioinformaticsLab/jpredapir/issues

JPred Server at: http://www.compbio.dundee.ac.uk/jpred/